+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Log files for tasks can now also be viewed with `logs <task-name>`
 
 ## [0.9.1-beta] - 2021-06-14
 ### Fixed